feat: Make some improvements to configure-zsh
Ensure the environment of the script is bash to avoid problems switching
the shell (Plasma hangs when the desktop loads otherwise).

Since Zsh can still be used as a shell without setting it as the user's
default, install oh-my-zsh with plugins regardless of whether it is the
default.

Enable the service earlier

Clean things up a bit